requirements.txt
numpy==2.2.6
sounddevice==0.5.2
openai-whisper==20240930
pydub==0.25.1
Flask==3.1.1